2012-10-17 03:55 UTC+0700 Andi Jahja <xharbour/AT/telkom/net/id>   * winmake/makefile.vc     ! applying highest warning level -W4 with safe-warnings disabled       (-wd4127 -wd4054 -wd4055). Temporarily applied to core stuff only.
  * source/rdd/dbcmd.c
  * source/rdd/dbcmd53.c
  * source/rdd/workarea.c
  * source/rdd/wacore.c
  * source/rdd/wafunc.c
  * source/rdd/dbf1.c
  * source/rdd/dbf1net.c
  * source/rdd/dbffpt/dbffpt1.c
  * source/rdd/dbffpt/redbffpt1.c
  * source/rdd/hbsix/sxcompr.c
  * source/rdd/hsx/hsx.c
  * source/rdd/usrrdd/usrrdd.c
    ! yet more MSVC warnings (W4) suppressed

    /*
    NOTE: Still, there are truncation on HB_PUT_LE_UINT32 and HB_PUT_BE_UINT32
    macro implementation used in some DBFFPT and Six Compression routines. I am
    not sure if these are critical, but logically speaking, truncation is
    usually not a good thing.
    */

+++ xharbour/source/rdd/dbf1.c
@@ -2767,7 +2767,7 @@ static HB_ERRCODE hb_dbfCreate( DBFAREAP pArea, LPDBOPENINFO pCreateInfo )
          pArea->lpdbOpenInfo = NULL;
          return HB_FAILURE;
       }
-      pArea->bLockType = hb_itemGetNI( pItem );
+      pArea->bLockType = ( BYTE ) hb_itemGetNI( pItem );
       if( pArea->bLockType == 0 )
       {
          pArea->bLockType = DB_DBFLOCK_CLIP;
@@ -3683,7 +3683,7 @@ static HB_ERRCODE hb_dbfNewArea( DBFAREAP pArea )
    {
       PHB_ITEM pItem = hb_itemPutNI( NULL, 0 );
       if( SELF_RDDINFO( SELF_RDDNODE( &pArea->area ), RDDI_TABLETYPE, 0, pItem ) == HB_SUCCESS )
-         pArea->bTableType = hb_itemGetNI( pItem );
+         pArea->bTableType = ( BYTE ) hb_itemGetNI( pItem );
       hb_itemRelease( pItem );
    }
 
@@ -3753,7 +3753,7 @@ static HB_ERRCODE hb_dbfOpen( DBFAREAP pArea, LPDBOPENINFO pOpenInfo )
          pArea->lpdbOpenInfo = NULL;
          return HB_FAILURE;
       }
-      pArea->bLockType = hb_itemGetNI( pItem );
+      pArea->bLockType = ( BYTE ) hb_itemGetNI( pItem );
       if( ! pArea->bLockType )
          pArea->bLockType = DB_DBFLOCK_CLIP;
    }
@@ -5770,7 +5770,7 @@ static HB_ERRCODE hb_dbfRddInfo( LPRDDNODE pRDD, USHORT uiIndex, ULONG ulConnect
             case DB_DBF_STD:        /* standard dBase/Clipper DBF file */
             case DB_DBF_VFP:        /* VFP DBF file */
             case DB_DBF_IV:         /* dBase IV DBF file */
-               pData->bTableType = iType;
+               pData->bTableType = ( BYTE ) iType;
          }
          break;
       }
@@ -5789,7 +5789,7 @@ static HB_ERRCODE hb_dbfRddInfo( LPRDDNODE pRDD, USHORT uiIndex, ULONG ulConnect
 #ifndef HB_LONG_LONG_OFF
             case DB_DBFLOCK_XHB64:
 #endif
-               pData->bLockType = ( int ) iScheme;
+               pData->bLockType = ( BYTE ) iScheme;
          }
          break;
       }
